unify wonder_mail and wonder_mail_mid
