+++ data/data_80DED44.s
@@ -1,10 +1,5 @@
         .section .rodata
 
-	.global gUnknown_80DED44
-gUnknown_80DED44: @ 80DED44
-        @ replacing .incbin "baserom.gba", 0x000ded44, 0x4
-        .4byte 0x01
-
 	.global gUnknown_80DED48
 gUnknown_80DED48: @ 80DED48
         @ replacing .incbin "baserom.gba", 0x000ded48, 0x18
@@ -1211,21 +1206,3 @@ gUnknown_80E1FDC: @ 80E1FDC
         .align 2,0
         .string "pksdir0\0"
         .align 2,0
-
-	.global gUnknown_80E1FF0
-gUnknown_80E1FF0: @ 80E1FF0
-        .byte 0x00, 0x00, 0x00, 0x00
-        .byte 0x03, 0x00, 0x00, 0x00
-        .byte 0x00, 0x00, 0x00, 0x00
-        .byte 0x00, 0x00, 0x00, 0x00
-        .byte 0x00, 0x00, 0x00, 0x00
-        .byte 0x00, 0x00, 0x00, 0x00
-
-	.global gUnknown_80E2008
-gUnknown_80E2008: @ 80E2008
-        .byte 0x00, 0x00, 0x00, 0x00
-        .byte 0x06, 0x00, 0x00, 0x00
-        .byte 0x02, 0x00, 0x02, 0x00
-        .byte 0x1a, 0x00, 0x04, 0x00
-        .byte 0x06, 0x00, 0x00, 0x00
-        .byte 0x00, 0x00, 0x00, 0x00
